[ANDROSDK-1296] Refactor internal factories
