+++ core/src/main/java/org/hisp/dhis/android/core/event/internal/EventWithLimitCallFactory.kt
@@ -69,7 +69,7 @@ class EventWithLimitCallFactory @Inject internal constructor(
 
         return Observable.create { emitter ->
             var successfulSync = true
-            val bundles = eventQueryBundleFactory.getEventQueryBundles(params)
+            val bundles = eventQueryBundleFactory.getQueries(params)
 
             for (bundle in bundles) {
