+++ core/src/main/java/org/hisp/dhis/android/core/trackedentity/internal/TrackedEntityInstanceQueryFactory.kt
@@ -38,7 +38,6 @@ internal class TrackedEntityInstanceQueryFactory @Inject constructor(
     programSettingsObjectRepository: ProgramSettingsObjectRepository,
     lastUpdatedManager: TrackedEntityInstanceLastUpdatedManager,
     commonHelper: TrackerQueryFactoryCommonHelper,
-    globalHelper: TrackedEntityInstanceQueryGlobalHelper,
-    perProgramHelper: TrackedEntityInstanceQueryPerProgramHelper
-) : TrackerQueryFactory<TeiQuery>(programStore, programSettingsObjectRepository, lastUpdatedManager, commonHelper, globalHelper, perProgramHelper)
+    internalFactory: TrackedEntityInstanceQueryInternalFactory
+) : TrackerQueryFactory<TeiQuery, TrackedEntityInstanceSync>(programStore, programSettingsObjectRepository, lastUpdatedManager, commonHelper, internalFactory)
